Microservice architecture:
1) Added three different entry points to compile services;
2) Re-structured project's layout to satisfy the task's requirements;
3) Other small changes and improvements.
