+++ cmd/notification/main.go
@@ -1,0 +1,11 @@
+package main
+
+import (
+	"log"
+)
+
+func main() {
+	if err := rootCmd.Execute(); err != nil {
+		log.Fatal(err)
+	}
+}
