+++ docs/notification/README.md
@@ -1,0 +1,3 @@
+# notification
+Simple event notification programme that leverages RabbitMQ message-brocker functionality written in Go.
+
