+++ docs/watcher/README.md
@@ -1,0 +1,2 @@
+# watcher
+Simple DB watcher programme that enqueues events for later notification.
