+++ internal/go-calendar/domain/errors/error.go
