+++ internal/go-calendar/domain/models/event.go
