+++ internal/interfaces/queue.go
@@ -1,0 +1,6 @@
+package interfaces
+
+type MessageQueuer interface {
+	ProduceMessages() error
+	EmmitMessages() error
+}
